Moved layout decode/encode logic into serializer
