+++ src/lib/utils.ts
@@ -2,8 +2,8 @@ export class Utils {
     // https://stackoverflow.com/questions/8495687/split-array-into-chunks
     static chunk = (arr: any[], chunkSize: number) => {
         return arr.reduce((all, one, i) => {
-            const ch = Math.floor(i / chunkSize); 
-            all[ch] = [].concat((all[ch] || []),one); 
+            const ch = Math.floor(i / chunkSize);
+            all[ch] = [].concat((all[ch] || []),one);
             return all;
          }, [])
     }
